+++ InfiniTAM/ITMLib/Trackers/CPU/ITMExtendedTracker_CPU.h
@@ -14,7 +14,7 @@ namespace ITMLib
 	public:
 		ITMExtendedTracker_CPU(Vector2i imgSize, TrackerIterationType *trackingRegime, int noHierarchyLevels,
 			float terminationThreshold, float failureDetectorThreshold, float viewFrustum_min, float viewFrustum_max, 
-			float tukeyCutOff, float framesToSkip, float framesToWeight, const ITMLowLevelEngine *lowLevelEngine);
+			int tukeyCutOff, int framesToSkip, int framesToWeight, const ITMLowLevelEngine *lowLevelEngine);
 		~ITMExtendedTracker_CPU(void);
 	};
 }
