+++ InfiniTAM/ITMLib/Trackers/CUDA/ITMExtendedTracker_CUDA.h
@@ -21,7 +21,7 @@ namespace ITMLib
 	public:
 		ITMExtendedTracker_CUDA(Vector2i imgSize, TrackerIterationType *trackingRegime, int noHierarchyLevels,
 			float terminationThreshold, float failureDetectorThreshold, float viewFrustum_min, float viewFrustum_max, 
-			float tukeyCutOff, float framesToSkip, float framesToWeight, const ITMLowLevelEngine *lowLevelEngine);
+			int tukeyCutOff, int framesToSkip, int framesToWeight, const ITMLowLevelEngine *lowLevelEngine);
 		~ITMExtendedTracker_CUDA(void);
 	};
 }
