+++ InfiniTAM/ITMLib/Trackers/Interface/ITMExtendedTracker.h
@@ -58,7 +58,7 @@ namespace ITMLib
 		int currentFrameNo;
 
 		float viewFrustum_min, viewFrustum_max;
-		float tukeyCutOff, framesToSkip, framesToWeight;
+		int tukeyCutOff, framesToSkip, framesToWeight;
 
 		virtual int ComputeGandH(float &f, float *nabla, float *hessian, Matrix4f approxInvPose) = 0;
 
@@ -72,7 +72,7 @@ namespace ITMLib
 
 		ITMExtendedTracker(Vector2i imgSize, TrackerIterationType *trackingRegime, int noHierarchyLevels,
 			float terminationThreshold, float failureDetectorThreshold, 
-			float viewFrustum_min, float viewFrustum_max, float tukeyCutOff, float framesToSkip, float framesToWeight,
+			float viewFrustum_min, float viewFrustum_max, int tukeyCutOff, int framesToSkip, int framesToWeight,
 			const ITMLowLevelEngine *lowLevelEngine, MemoryDeviceType memoryType);
 		virtual ~ITMExtendedTracker(void);
 	};
